move module to root, is not a module anymore
