+++ params/internal/keeper/keeper.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 
 	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/cosmos/cosmos-sdk/params/internal/types"
+	"github.com/cosmos/cosmos-sdk/params/subspace"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/params/internal/types"
-	"github.com/cosmos/cosmos-sdk/x/params/subspace"
 
 	"github.com/tendermint/tendermint/libs/log"
 )
