+++ params/simulation/operations.go
@@ -3,9 +3,9 @@ package simulation
 import (
 	"math/rand"
 
+	"github.com/cosmos/cosmos-sdk/params/internal/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-	"github.com/cosmos/cosmos-sdk/x/params/internal/types"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
 )
