+++ params/spec/02_subspace.md
