+++ params/subspace/table_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/cosmos/cosmos-sdk/x/params/subspace"
+	"github.com/cosmos/cosmos-sdk/params/subspace"
 )
 
 func TestKeyTable(t *testing.T) {
