+++ x/distribution/types/params.go
@@ -5,8 +5,8 @@ import (
 
 	"gopkg.in/yaml.v2"
 
+	"github.com/cosmos/cosmos-sdk/params"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/params"
 )
 
 const (
