+++ x/evidence/internal/types/params.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/cosmos/cosmos-sdk/x/params"
+	"github.com/cosmos/cosmos-sdk/params"
 
 	"gopkg.in/yaml.v2"
 )
