+++ x/slashing/internal/types/expected_keepers.go
@@ -3,9 +3,9 @@
 package types
 
 import (
+	"github.com/cosmos/cosmos-sdk/params"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authexported "github.com/cosmos/cosmos-sdk/x/auth/exported"
-	"github.com/cosmos/cosmos-sdk/x/params"
 	stakingexported "github.com/cosmos/cosmos-sdk/x/staking/exported"
 )
