+++ x/staking/keeper/params.go
@@ -3,8 +3,8 @@ package keeper
 import (
 	"time"
 
+	"github.com/cosmos/cosmos-sdk/params"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/params"
 	"github.com/cosmos/cosmos-sdk/x/staking/types"
 )
